issue #189: move routes into main.py
